+++ src/main/java/org/scion/jpan/ScionUtil.java
@@ -44,17 +44,21 @@ public class ScionUtil {
     if (parts.length != 2) {
       throw new IllegalArgumentException("invalid ISD-AS: value=" + ia);
     }
-    int isd = Integer.parseUnsignedInt(parts[0], 10);
+    int isd = parseISD(parts[0]);
     long as = parseAS(parts[1]);
     checkLimits(isd, as);
     return Integer.toUnsignedLong(isd) << AS_BITS | (as & MAX_AS);
+  }
+
+  public static int parseISD(String isd) {
+    return Integer.parseUnsignedInt(isd, 10);
   }
 
   /**
    * ParseAS parses an AS from a decimal (in the case of the 32bit BGP AS number space) or
    * ipv6-style hex (in the case of SCION-only AS numbers) string.
    */
-  private static long parseAS(String as) {
+  public static long parseAS(String as) {
     String[] parts = as.split(":");
     if (parts.length == 1) {
       // Must be a BGP AS, parse as 32-bit decimal number