+++ CHANGELOG.md
@@ -50,6 +50,8 @@ For example: `Path.getFirstHopAddress()`, `DatagramChannel.setPathPolicy()`
 - Added keep-alive protocol for NAT. [#151](https://github.com/scionproto-contrib/jpan/pull/151)
 - Added implementation of STUN responder (currently not needed)
   [#154](https://github.com/scionproto-contrib/jpan/pull/154)
+- `PathPolicy` created from Path Policy Language
+  [#158](https://github.com/scionproto-contrib/jpan/pull/158)
 
 ### Changed